Методы для издателя перенесены в компонент Publisher в папке API
