+++ react-web-client/src/redux/state.js
@@ -1,6 +1,7 @@
 import * as axios from "axios";
 import { getPainterStyle } from "./API/PainterStyles";
 import { getPainterByCount } from "./API/Painter";
+import { getPublishersByCount } from "./API/Publisher";
 
 let state = {
     contentPage: {
@@ -35,60 +36,8 @@ export let initializeData = () => {
     getInterpretersByCount(10);
     getAuthorByCount(10);
     getBooksByCount(10);
-    // rerenderEntireTree(state);
 }
 
-
-
-// TODO: вынести в отдельный компонент
-export let addPublisher = (name) => {
-    let publisher = {
-        id: newGuid(),
-        name,
-        booksIds: [
-            newGuid()
-        ]
-    }
-    axios.post('https://localhost:44394/api/Publisher', publisher).then(response => {
-        console.log('add publisher', response);
-    })
-
-    getPublishersByCount(10);
-    // rerenderEntireTree(state);
-}
-
-export let deletePublisher = (name) => {
-    let publisherId = state.contentPage.publishers.find(ps => ps.name === name).id
-
-    axios.delete('https://localhost:44394/api/Publisher/' + publisherId).then(response => {
-        console.log('delete publisher', response);
-    })
-    initializeData();
-}
-
-export let getPublishersByCount = (count) => {
-    axios.get('https://localhost:44394/api/Publisher/GetPublishers/take/' + count + '/skip/0').then(response => {
-        console.log('get publisher', response);
-        state.contentPage.publishers = response.data.previewPublishers;
-    })
-    return state.contentPage.publishers;
-}
-
-export let updatePublisher = (oldName, name) => {
-    let publisherId = state.contentPage.publishers.find(ps => ps.name === oldName).id
-    let publisher = {
-        id: publisherId,
-        name,
-        booksIds: [
-            newGuid()
-        ]
-    }
-    axios.put('https://localhost:44394/api/Publisher/', publisher).then(response => {
-        console.log('update publisher', response);
-    })
-
-    initializeData();
-}
 
 // TODO: вынести в отдельный компонент
 export let getCoverTypes = () => {
