+++ react-web-client/src/render.js
@@ -5,9 +5,9 @@ import App from './App';
 import * as serviceWorker from './serviceWorker';
 import { getPainterStyle, deletePainterStyle, updatePainterStyle, setPainterStyle } from './redux/API/PainterStyles';
 import { setPainter, deletePainter, updatePainter, getPainterByCount } from './redux/API/Painter';
+import { addPublisher, getPublishersByCount, deletePublisher, updatePublisher } from './redux/API/Publisher';
 
 import {
-  addPublisher, getPublishersByCount, deletePublisher, updatePublisher,
   getCoverTypes, addCoverType, updateCoverType, deleteCoverType,
   addInterpreter, deleteInterpreter, updateInterpreter,
   addAuthor, updateAuthor, deleteAuthor,
